Tratamentos de erro Charge Not Found

+++ src/modules/charge/useCases/delete-charge.usecase.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { ChargeRepository } from "../repositories/charge.repository";
 
 
@@ -7,6 +7,11 @@ export class DeleteChargeUseCase {
     constructor(private readonly chargeRepository: ChargeRepository) {}
 
     async deleteCharge(id: string) {
+        const chargeExist = await this.chargeRepository.findOne(id);
+        if (!chargeExist) {
+            throw new NotFoundException("Charge not found.")
+        }
+        
         return await this.chargeRepository.deleteById(id);
     }
 }
